feat: fixing roteiro5 issues

+++ src/classes/BinOp.py
@@ -6,12 +6,12 @@ class BinOp(Node):
         node1 = self.children[0].Evaluate()
         node2 = self.children[1].Evaluate()
         if self.variant == "+":
-            return node1+node2
+            return int(node1+node2)
         elif self.variant == "-":
-            return node1-node2
+            return int(node1-node2)
         elif self.variant == "*":
-            return node1*node2
+            return int(node1*node2)
         elif self.variant == "/":
-            return node1/node2
+            return int(node1/node2)
         else:
             raise Exception("Invalid token")
